backend/requirements.txt
fastapi
uvicorn[standard]
pykrx
finance-datareader
pandas
numpy
supabase
python-dotenv
